implemented the changes mentioned by Gabriel concerning the tutorial.
